Combining the TopBar with the HomeScreen in order that it works together
Remove old top bar
