+++ app/src/main/java/com/github/swent/echo/compose/components/HomeScreen.kt
@@ -1,7 +1,13 @@
 package com.github.swent.echo.compose.components
 
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
 import androidx.compose.material.icons.Icons
@@ -33,12 +39,30 @@ import androidx.compose.material3.ModalNavigationDrawer
 import androidx.compose.material3.NavigationDrawerItem
 import androidx.compose.material3.NavigationDrawerItemDefaults
 import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBarDefaults
+import androidx.compose.material3.rememberDrawerState
+import androidx.compose.material3.rememberTopAppBarState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.MutableState
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableIntStateOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.platform.testTag
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import com.github.swent.echo.R
 import com.github.swent.echo.compose.map.MAP_CENTER
 import com.github.swent.echo.compose.map.MapDrawer
 import com.github.swent.echo.data.model.Event
@@ -46,6 +70,24 @@ import com.github.swent.echo.data.model.Location
 import com.github.swent.echo.ui.navigation.NavigationActions
 import java.time.Instant
 import java.util.Date
+import kotlinx.coroutines.launch
+
+/**
+ * data class to display the navigation items in the hamburger menu.
+ *
+ * @param title the title of the navigation item
+ * @param selectedIcon the icon to display when the item is selected
+ * @param unselectedIcon the icon to display when the item is not selected
+ * @param badgeCount the count to display as a badge on the item (for example friends requests)
+ * @param navOnClick the action to perform when the item is clicked
+ */
+data class NavigationItem(
+    val title: String,
+    val selectedIcon: ImageVector,
+    val unselectedIcon: ImageVector,
+    val badgeCount: Int? = null,
+    val navOnClick: (() -> Unit)? = {}
+)
 
 enum class Overlay {
     NONE,
@@ -71,20 +113,63 @@ data class DisplayEventInfo(
  * the hamburger menu when the button is clicked on the top bar, the search bottom sheet, the list
  * view, and the map.
  */
+@OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun HomeScreen(navActions: NavigationActions) {
+    // State for the MapOrListMode
     val overlay = remember { mutableStateOf(Overlay.NONE) }
     val mode = remember { mutableStateOf(MapOrListMode.MAP) }
     val displayEventInfo = remember { mutableStateOf<DisplayEventInfo?>(null) }
-    Scaffold(
-        modifier = Modifier.testTag("home_screen"),
-        topBar = {
-            TopBar(navActions) {
-                if (mode.value == MapOrListMode.MAP) {
-                    mode.value = MapOrListMode.LIST
-                } else {
-                    mode.value = MapOrListMode.MAP
-                }
+
+    // Scroll behavior for the top app bar, makes it pinned
+    val scrollBehavior = TopAppBarDefaults.pinnedScrollBehavior(rememberTopAppBarState())
+
+    // Drawer state to open and close the hamburger menu
+    val drawerState = rememberDrawerState(initialValue = DrawerValue.Closed)
+    val scope = rememberCoroutineScope()
+    var selectedItemIndex by rememberSaveable { mutableIntStateOf(0) }
+
+    // List of navigation items to display in the hamburger menu
+    // TODO: Change the icons to be more meaningful when we have the google icons
+    // TODO: Add a navOnClick to the item to navigate to the corresponding screen as soon as you
+    // implement it
+    val items =
+        listOf(
+            NavigationItem(
+                title = stringResource(id = R.string.hamburger_my_profile),
+                selectedIcon = Icons.Filled.Person,
+                unselectedIcon = Icons.Outlined.Person
+            ),
+            NavigationItem(
+                title = stringResource(id = R.string.hamburger_my_events),
+                selectedIcon = Icons.Filled.DateRange,
+                unselectedIcon = Icons.Outlined.DateRange,
+            ),
+            NavigationItem(
+                title = stringResource(id = R.string.hamburger_friends),
+                selectedIcon = Icons.Filled.Face,
+                unselectedIcon = Icons.Outlined.Face,
+                badgeCount = 3
+            ),
+            NavigationItem(
+                title = stringResource(id = R.string.hamburger_settings),
+                selectedIcon = Icons.Filled.Settings,
+                unselectedIcon = Icons.Outlined.Settings,
+            ),
+            NavigationItem(
+                title = stringResource(id = R.string.hamburger_create_event),
+                selectedIcon = Icons.Filled.AddCircle,
+                unselectedIcon = Icons.Outlined.AddCircle,
+            ),
+            NavigationItem(
+                title = stringResource(id = R.string.hamburger_add_friends),
+                selectedIcon = Icons.Filled.Add,
+                unselectedIcon = Icons.Outlined.Add,
+            ),
+            NavigationItem(
+                title = stringResource(id = R.string.hamburger_help),
+                selectedIcon = Icons.Filled.Build,
+                unselectedIcon = Icons.Outlined.Build,
             )
         )
 
@@ -96,16 +181,16 @@ fun HomeScreen(navActions: NavigationActions) {
                 // Profile picture, name and class
                 Box(
                     modifier =
-                    Modifier.background(MaterialTheme.colorScheme.primaryContainer)
-                        .fillMaxWidth()
-                        .padding(NavigationDrawerItemDefaults.ItemPadding)
-                        .testTag("profile_box")
+                        Modifier.background(MaterialTheme.colorScheme.primaryContainer)
+                            .fillMaxWidth()
+                            .padding(NavigationDrawerItemDefaults.ItemPadding)
+                            .testTag("profile_box")
                 ) {
                     Column(
                         modifier =
-                        Modifier.align(Alignment.TopStart)
-                            .padding(8.dp)
-                            .testTag("profile_sheet")
+                            Modifier.align(Alignment.TopStart)
+                                .padding(8.dp)
+                                .testTag("profile_sheet")
                     ) {
                         // TODO: Replace with actual profile picture
                         Image(
@@ -134,9 +219,9 @@ fun HomeScreen(navActions: NavigationActions) {
                             scope.launch { drawerState.close() }
                         },
                         modifier =
-                        Modifier.align(Alignment.TopEnd)
-                            .padding(8.dp)
-                            .testTag("close_button_hamburger_menu")
+                            Modifier.align(Alignment.TopEnd)
+                                .padding(8.dp)
+                                .testTag("close_button_hamburger_menu")
                     ) {
                         Icon(
                             imageVector = Icons.Filled.Close,
@@ -158,11 +243,11 @@ fun HomeScreen(navActions: NavigationActions) {
                         icon = {
                             Icon(
                                 imageVector =
-                                if (index == selectedItemIndex) {
-                                    item.selectedIcon
-                                } else {
-                                    item.unselectedIcon
-                                },
+                                    if (index == selectedItemIndex) {
+                                        item.selectedIcon
+                                    } else {
+                                        item.unselectedIcon
+                                    },
                                 contentDescription = item.title
                             )
                         },
@@ -170,15 +255,71 @@ fun HomeScreen(navActions: NavigationActions) {
                             item.badgeCount?.let { Text(text = item.badgeCount.toString()) }
                         },
                         modifier =
-                        Modifier.padding(NavigationDrawerItemDefaults.ItemPadding)
-                            .testTag("navigation_item_$index")
+                            Modifier.padding(NavigationDrawerItemDefaults.ItemPadding)
+                                .testTag("navigation_item_$index")
                     )
                 }
             }
         },
-        floatingActionButton = { SearchButton(onClick = { overlay.value = Overlay.SEARCH_SHEET }) }
-    ) { paddingValues ->
-        Content(paddingValues, overlay, mode, navActions, displayEventInfo)
+        drawerState = drawerState,
+        modifier = Modifier.testTag("hamburger_menu"),
+        gesturesEnabled = false
+    ) {
+        Scaffold(
+            modifier = Modifier.testTag("home_screen"),
+            topBar = {
+                CenterAlignedTopAppBar(
+                    colors =
+                        TopAppBarDefaults.topAppBarColors(
+                            containerColor = MaterialTheme.colorScheme.primaryContainer,
+                            titleContentColor = MaterialTheme.colorScheme.onPrimary,
+                        ),
+                    title = {
+                        Text(
+                            "Echo",
+                            maxLines = 1,
+                            overflow = TextOverflow.Ellipsis, // should not happen
+                            fontWeight = FontWeight.SemiBold
+                        )
+                    },
+                    navigationIcon = { // hamburger menu
+                        IconButton(
+                            onClick = { scope.launch { drawerState.open() } },
+                            modifier = Modifier.testTag("menu_button")
+                        ) {
+                            Icon(
+                                imageVector = Icons.Filled.Menu,
+                                contentDescription = "Menu to access all subpages"
+                            )
+                        }
+                    },
+                    actions = { // search icon
+                        IconButton(
+                            onClick = {
+                                if (mode.value == MapOrListMode.MAP) {
+                                    mode.value = MapOrListMode.LIST
+                                } else {
+                                    mode.value = MapOrListMode.MAP
+                                }
+                            },
+                            modifier = Modifier.testTag("list_map_mode_button")
+                        ) {
+                            Icon(
+                                imageVector = Icons.Filled.List,
+                                contentDescription = "Search icon to access the search screen"
+                            )
+                        }
+                    },
+                    scrollBehavior = scrollBehavior,
+                    modifier = Modifier.testTag("top_bar")
+                )
+            },
+            floatingActionButton = {
+                SearchButton(onClick = { overlay.value = Overlay.SEARCH_SHEET })
+            }
+        ) { paddingValues ->
+            Content(paddingValues, overlay, mode, navActions, displayEventInfo)
+        }
     }
 }
 
@@ -219,6 +360,7 @@ private fun Content(
                         Instant.now(),
                     ),
                 tags = emptySet(),
+                creatorId = "d"
             ),
             Event(
                 eventId = "b",
@@ -243,6 +385,7 @@ private fun Content(
                         Instant.now(),
                     ),
                 tags = emptySet(),
+                creatorId = "e"
             )
         )
